+++ Assets/AppCenter/Plugins/AppCenterSDK/Distribute/Android/DistributeDelegate.cs
@@ -27,6 +27,11 @@ namespace Microsoft.AppCenter.Unity.Distribute.Internal
             var releaseDetails = ReleaseDetailsHelper.ReleaseDetailsConvert(details);
             return Distribute.ReleaseAvailable.Invoke(releaseDetails);
         }
+
+        void onNoReleaseAvailable(AndroidJavaObject activity)
+        {
+            Distribute.NoReleaseAvailable?.Invoke();
+        }
     }
 }
 #endif