+++ Assets/AppCenter/Plugins/AppCenterSDK/Distribute/iOS/DistributeInternal.cs
@@ -29,6 +29,15 @@ namespace Microsoft.AppCenter.Unity.Distribute.Internal
             var releaseDetails = ReleaseDetailsHelper.ReleaseDetailsConvert(details);
             return Distribute.ReleaseAvailable.Invoke(releaseDetails);
         }
+
+        delegate void NoReleaseAvailableDelegate();
+
+        static NoReleaseAvailableDelegate noReleaseDel;
+        [MonoPInvokeCallback(typeof(NoReleaseAvailableDelegate))]
+        static void NoReleaseAvailableFunc()
+        {
+            Distribute.NoReleaseAvailable?.Invoke();
+        }
 #endregion
 
         public static void PrepareEventHandlers()
@@ -46,7 +55,14 @@ namespace Microsoft.AppCenter.Unity.Distribute.Internal
         {
             appcenter_unity_distribute_set_delegate();
             del = ReleaseAvailableFunc;
+            noReleaseDel = NoReleaseAvailableFunc;
             appcenter_unity_distribute_set_release_available_impl(del);
+            appcenter_unity_distribute_set_no_release_available_impl(noReleaseDel);
+        }
+
+        public static void StartDistribute()
+        {
+            appcenter_unity_start_distribute();
         }
 
         public static void AddNativeType(List<IntPtr> nativeTypes)
@@ -110,6 +126,9 @@ namespace Microsoft.AppCenter.Unity.Distribute.Internal
         private static extern void appcenter_unity_distribute_set_release_available_impl(ReleaseAvailableDelegate functionPtr);
 
         [DllImport("__Internal")]
+        private static extern void appcenter_unity_distribute_set_no_release_available_impl(NoReleaseAvailableDelegate functionPtr);
+
+        [DllImport("__Internal")]
         private static extern void appcenter_unity_distribute_replay_release_available();
 
         [DllImport("__Internal")]
@@ -117,6 +136,9 @@ namespace Microsoft.AppCenter.Unity.Distribute.Internal
 
         [DllImport("__Internal")]
         private static extern void appcenter_unity_distribute_check_for_update();
+
+        [DllImport("__Internal")]
+        private static extern void appcenter_unity_start_distribute();
 
 #endregion
     }