+++ Assets/AppCenter/Plugins/iOS/Distribute/DistributeDelegate.m
@@ -9,6 +9,7 @@
 @property MSACReleaseDetails *unprocessedDetails;
 @property NSObject *lockObject;
 @property ReleaseAvailableFunction releaseAvailableHandler;
+@property NoReleaseAvailableFunction noReleaseAvailableHandler;
 
 @end
 
@@ -39,6 +40,12 @@
   }
 }
 
+- (void)setNoReleaseAvailableImplementation:(NoReleaseAvailableFunction)implementation {
+  @synchronized (_lockObject) {
+    _noReleaseAvailableHandler = implementation;
+  }
+}
+
 - (BOOL)distribute:(MSACDistribute *)distribute releaseAvailableWithDetails:(MSACReleaseDetails *)details {
   ReleaseAvailableFunction handlerCopy = nil;
   @synchronized (_lockObject) {
@@ -50,7 +57,7 @@
   }
 }
 
-- (void) replayReleaseAvailable {
+- (void)replayReleaseAvailable {
   ReleaseAvailableFunction handlerCopy = nil;
   MSACReleaseDetails *unprocessedCopy = _unprocessedDetails;
   @synchronized (_lockObject) {
@@ -59,6 +66,17 @@
   }
   if (unprocessedCopy && handlerCopy) {
     handlerCopy(unprocessedCopy);
+  }
+}
+
+- (void)distributeNoReleaseAvailable:(MSACDistribute *)distribute {
+  NoReleaseAvailableFunction handlerCopy = nil;
+  @synchronized (_lockObject) {
+    handlerCopy = _noReleaseAvailableHandler;
+    if (handlerCopy) {
+      handlerCopy();
+    }
+    return;
   }
 }