+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # App Center SDK for Unity Change Log
 
+## Version 4.0.1 (under development)
+
+### App Center Distribute
+
+#### iOS/Android
+
+* **[Feature]** Add `NoReleaseAvailable` callback to distribute listener.
+* **[Fix]** Fix show the custom dialog update after the application start.
+
 ## Release 4.0.0
 
 ### App Center